+++ taxcalc/utils.py
@@ -1,5 +1,5 @@
 """
-PUBLIC utility functions for Tax-Calculator.
+PUBLIC low-level utility functions for Tax-Calculator.
 """
 # CODING-STYLE CHECKS:
 # pep8 --ignore=E402 utils.py
@@ -28,14 +28,15 @@ from taxcalc.utilsprvt import (weighted_count_lt_zero,
                                EPSILON)
 
 
-STATS_COLUMNS = ['expanded_income', 'c00100', 'aftertax_income', 'standard',
-                 'c04470', 'c04600', 'c04800', 'taxbc', 'c62100', 'c09600',
-                 'c05800', 'othertaxes', 'refund', 'c07100', 'iitax',
-                 'payrolltax', 'combined', 's006']
-
 # Items in the DIST_TABLE_COLUMNS list below correspond to the items in the
 # DIST_TABLE_LABELS list below; this correspondence allows us to use this
 # labels list to map a label to the correct column in a distribution table.
+
+DIST_VARIABLES = ['expanded_income', 'c00100', 'aftertax_income', 'standard',
+                  'c04470', 'c04600', 'c04800', 'taxbc', 'c62100', 'c09600',
+                  'c05800', 'othertaxes', 'refund', 'c07100', 'surtax',
+                  'iitax', 'payrolltax', 'combined', 's006']
+
 DIST_TABLE_COLUMNS = ['s006',
                       'c00100',
                       'num_returns_StandardDed',
@@ -83,6 +84,10 @@ DIST_TABLE_LABELS = ['Returns',
 # Items in the DIFF_TABLE_COLUMNS list below correspond to the items in the
 # DIFF_TABLE_LABELS list below; this correspondence allows us to use this
 # labels list to map a label to the correct column in a difference table.
+
+DIFF_VARIABLES = ['expanded_income', 'c00100', 'aftertax_income',
+                  'iitax', 'payrolltax', 'combined', 's006']
+
 DIFF_TABLE_COLUMNS = ['count',
                       'tax_cut',
                       'perc_cut',
@@ -241,16 +246,16 @@ def get_sums(pdf):
     return pd.Series(sums, name='sums')
 
 
-def create_distribution_table(obj, groupby, income_measure, result_type):
+def create_distribution_table(vdf, groupby, income_measure, result_type):
     """
-    Get results from object, sort them based on groupby using income_measure,
+    Get results from vdf, sort them based on groupby using income_measure,
     manipulate them based on result_type, and return them as a table.
 
     Parameters
     ----------
-    obj : any object with array-like attributes named as in STATS_COLUMNS list
-        Examples include a Tax-Calculator Calculator object and a
-        Pandas DataFrame object.
+    vdf : Pandas DataFrame including columns named as in STATS_COLUMNS list
+        for example, object returned from Calculator dataframe method in a
+        call like this: vdf = calc.dataframe(STATS_COLUMNS)
 
     groupby : String object
         options for input: 'weighted_deciles', 'webapp_income_bins',
@@ -326,21 +331,19 @@ def create_distribution_table(obj, groupby, income_measure, result_type):
         return sdf
 
     # main logic of create_distribution_table
-    if result_type != 'weighted_sum' and result_type != 'weighted_avg':
-        msg = "result_type must be either 'weighted_sum' or 'weighted_avg'"
-        raise ValueError(msg)
+    assert isinstance(vdf, pd.DataFrame)
+    assert (groupby == 'weighted_deciles' or
+            groupby == 'webapp_income_bins' or
+            groupby == 'large_income_bins' or
+            groupby == 'small_income_bins')
+    assert result_type == 'weighted_sum' or result_type == 'weighted_avg'
     assert (income_measure == 'expanded_income' or
-            income_measure == 'c00100' or
             income_measure == 'expanded_income_baseline' or
+            income_measure == 'c00100' or
             income_measure == 'c00100_baseline')
-    if income_measure in STATS_COLUMNS:
-        columns = STATS_COLUMNS
-    else:
-        columns = STATS_COLUMNS + [income_measure]
-    if isinstance(obj, pd.DataFrame):
-        res = copy.deepcopy(obj)
-    else:
-        res = obj.dataframe(columns)
+    assert income_measure in vdf
+    # copy vdf and add variable columns
+    res = copy.deepcopy(vdf)
     res = add_columns(res)
     # sort the data given specified groupby and income_measure
     if groupby == 'weighted_deciles':
@@ -351,11 +354,6 @@ def create_distribution_table(obj, groupby, income_measure, result_type):
         pdf = add_income_bins(res, income_measure, bin_type='tpc')
     elif groupby == 'small_income_bins':
         pdf = add_income_bins(res, income_measure, bin_type='soi')
-    else:
-        msg = ("groupby must be either 'weighted_deciles' or "
-               "'webapp_income_bins' or 'large_income_bins' or "
-               "'small_income_bins'")
-        raise ValueError(msg)
     # construct weighted_sum table
     gpdf = pdf.groupby('bins', as_index=False)
     dist_table = stat_dataframe(gpdf)
@@ -384,18 +382,20 @@ def create_distribution_table(obj, groupby, income_measure, result_type):
     return dist_table
 
 
-def create_difference_table(obj1, obj2, groupby, income_measure, tax_to_diff):
+def create_difference_table(vdf1, vdf2, groupby, income_measure, tax_to_diff):
     """
-    Get results from two different obj, construct tax difference results,
+    Get results from two different vdf, construct tax difference results,
     and return the difference statistics as a table.
 
     Parameters
     ----------
-    obj1 : baseline object is either a Tax-Calculator Calculator object or
-           a Pandas DataFrame including columns in STATS_COLUMNS list
+    vdf1 : Pandas DataFrame object including columns in the DIFF_VARIABLES
+           list drawn from a baseline Calculator object using the
+           Calculator.dataframe method
 
-    obj2 : reform object is either a Tax-Calculator Calculator object or
-           a Pandas DataFrame including columns in STATS_COLUMNS list
+    vdf2 : Pandas DataFrame object including columns in the DIFF_VARIABLES
+           list drawn from a baseline Calculator object using the
+           Calculator.dataframe method
 
     groupby : String object
         options for input: 'weighted_deciles', 'webapp_income_bins',
@@ -462,6 +462,7 @@ def create_difference_table(obj1, obj2, groupby, income_measure, tax_to_diff):
             sdf['atinc1'] = gpdf.apply(weighted_sum, 'atinc1')
             sdf['atinc2'] = gpdf.apply(weighted_sum, 'atinc2')
             return sdf
+
         # main logic of diff_table_stats function
         # add bin column to res2 given specified groupby and income_measure
         if groupby == 'weighted_deciles':
@@ -472,11 +473,6 @@ def create_difference_table(obj1, obj2, groupby, income_measure, tax_to_diff):
             pdf = add_income_bins(res2, income_measure, bin_type='tpc')
         elif groupby == 'small_income_bins':
             pdf = add_income_bins(res2, income_measure, bin_type='soi')
-        else:
-            msg = ("groupby must be either "
-                   "'weighted_deciles' or 'webapp_income_bins' "
-                   "or 'large_income_bins' or 'small_income_bins'")
-            raise ValueError(msg)
         # create grouped Pandas DataFrame
         gpdf = pdf.groupby('bins', as_index=False)
         # create difference table statistics from gpdf in a new DataFrame
@@ -504,17 +500,20 @@ def create_difference_table(obj1, obj2, groupby, income_measure, tax_to_diff):
             diffs = diffs.append(sdf, ignore_index=True)
         return diffs
     # main logic of create_difference_table
-    is_dframe1 = isinstance(obj1, pd.DataFrame)
-    is_dframe2 = isinstance(obj2, pd.DataFrame)
-    assert is_dframe1 == is_dframe2
-    if is_dframe1:
-        res1 = copy.deepcopy(obj1)
-        res2 = copy.deepcopy(obj2)
-    else:
-        assert obj1.current_year == obj2.current_year
-        res1 = obj1.dataframe(STATS_COLUMNS)
-        res2 = obj2.dataframe(STATS_COLUMNS)
-    assert income_measure == 'expanded_income' or income_measure == 'c00100'
+    assert isinstance(vdf1, pd.DataFrame)
+    assert isinstance(vdf2, pd.DataFrame)
+    assert (groupby == 'weighted_deciles' or
+            groupby == 'webapp_income_bins' or
+            groupby == 'large_income_bins' or
+            groupby == 'small_income_bins')
+    assert (income_measure == 'expanded_income' or
+            income_measure == 'c00100')
+    assert income_measure in vdf1
+    assert (tax_to_diff == 'iitax' or
+            tax_to_diff == 'payrolltax' or
+            tax_to_diff == 'combined')
+    res1 = copy.deepcopy(vdf1)
+    res2 = copy.deepcopy(vdf2)
     baseline_income_measure = income_measure + '_baseline'
     res2[baseline_income_measure] = res1[income_measure]
     res2['tax_diff'] = res2[tax_to_diff] - res1[tax_to_diff]
@@ -536,27 +535,31 @@ def create_difference_table(obj1, obj2, groupby, income_measure, tax_to_diff):
     return diffs
 
 
-def create_diagnostic_table(calc):
+def create_diagnostic_table(vdf, year):
     """
-    Extract diagnostic table from specified Calculator object.
-    This function leaves the specified calc object unchanged.
+    Extract single-year diagnostic table from Pandas DataFrame object
+    derived from a Calculator object using the dataframe(DIST_VARIABLES)
+    method.
 
     Parameters
     ----------
-    calc : Calculator class object
+    vdf : Pandas DataFrame object containing the variables
+
+    year : calendar year for which variables were drawn from Calculator object
 
     Returns
     -------
-    Pandas DataFrame object containing the table for calc.current_year
+    Pandas DataFrame object containing the diagnostic table
     """
+    # pylint: disable=too-many-statements
     def diagnostic_table_odict(recs):
         """
         Nested function that extracts diagnostic table dictionary from
-        the specified Records object, recs.
+        the specified Pandas DataFrame object, vdf.
 
         Parameters
         ----------
-        recs : Records class object
+        vdf : Pandas DataFrame object containing the variables
 
         Returns
         -------
@@ -567,113 +570,80 @@ def create_diagnostic_table(calc):
         in_billions = 1.0e-9
         odict = collections.OrderedDict()
         # total number of filing units
-        odict['Returns (#m)'] = recs.s006.sum() * in_millions
+        wghts = vdf['s006']
+        odict['Returns (#m)'] = wghts.sum() * in_millions
         # adjusted gross income
-        odict['AGI ($b)'] = (recs.c00100 * recs.s006).sum() * in_billions
+        agi = vdf['c00100']
+        odict['AGI ($b)'] = (agi * wghts).sum() * in_billions
         # number of itemizers
-        num = (recs.s006[(recs.c04470 > 0.) * (recs.c00100 > 0.)].sum())
+        num = (wghts[(vdf['c04470'] > 0.) & (agi > 0.)].sum())
         odict['Itemizers (#m)'] = num * in_millions
         # itemized deduction
-        ided1 = recs.c04470 * recs.s006
-        val = ided1[recs.c04470 > 0.].sum()
+        ided1 = vdf['c04470'] * wghts
+        val = ided1[vdf['c04470'] > 0.].sum()
         odict['Itemized Deduction ($b)'] = val * in_billions
         # number of standard deductions
-        num = recs.s006[(recs.standard > 0.) * (recs.c00100 > 0.)].sum()
+        num = wghts[(vdf['standard'] > 0.) & (agi > 0.)].sum()
         odict['Standard Deduction Filers (#m)'] = num * in_millions
         # standard deduction
-        sded1 = recs.standard * recs.s006
-        val = sded1[(recs.standard > 0.) * (recs.c00100 > 0.)].sum()
+        sded1 = recs.standard * wghts
+        val = sded1[(vdf['standard'] > 0.) & (agi > 0.)].sum()
         odict['Standard Deduction ($b)'] = val * in_billions
         # personal exemption
-        val = (recs.c04600 * recs.s006)[recs.c00100 > 0.].sum()
+        val = (vdf['c04600'] * wghts)[agi > 0.].sum()
         odict['Personal Exemption ($b)'] = val * in_billions
         # taxable income
-        val = (recs.c04800 * recs.s006).sum()
+        val = (vdf['c04800'] * wghts).sum()
         odict['Taxable Income ($b)'] = val * in_billions
         # regular tax liability
-        val = (recs.taxbc * recs.s006).sum()
+        val = (vdf['taxbc'] * wghts).sum()
         odict['Regular Tax ($b)'] = val * in_billions
         # AMT taxable income
-        odict['AMT Income ($b)'] = ((recs.c62100 * recs.s006).sum() *
+        odict['AMT Income ($b)'] = ((vdf['c62100'] * wghts).sum() *
                                     in_billions)
         # total AMT liability
-        odict['AMT Liability ($b)'] = ((recs.c09600 * recs.s006).sum() *
+        odict['AMT Liability ($b)'] = ((vdf['c09600'] * wghts).sum() *
                                        in_billions)
         # number of people paying AMT
-        odict['AMT Filers (#m)'] = (recs.s006[recs.c09600 > 0.].sum() *
+        odict['AMT Filers (#m)'] = (wghts[vdf['c09600'] > 0.].sum() *
                                     in_millions)
         # tax before credits
-        val = (recs.c05800 * recs.s006).sum()
+        val = (vdf['c05800'] * wghts).sum()
         odict['Tax before Credits ($b)'] = val * in_billions
         # refundable credits
-        val = (recs.refund * recs.s006).sum()
+        val = (vdf['refund'] * wghts).sum()
         odict['Refundable Credits ($b)'] = val * in_billions
         # nonrefundable credits
-        val = (recs.c07100 * recs.s006).sum()
+        val = (vdf['c07100'] * wghts).sum()
         odict['Nonrefundable Credits ($b)'] = val * in_billions
         # reform surtaxes (part of federal individual income tax liability)
-        val = (recs.surtax * recs.s006).sum()
+        val = (vdf['surtax'] * wghts).sum()
         odict['Reform Surtaxes ($b)'] = val * in_billions
         # other taxes on Form 1040
-        val = (recs.othertaxes * recs.s006).sum()
+        val = (vdf['othertaxes'] * wghts).sum()
         odict['Other Taxes ($b)'] = val * in_billions
         # federal individual income tax liability
-        val = (recs.iitax * recs.s006).sum()
+        val = (vdf['iitax'] * wghts).sum()
         odict['Ind Income Tax ($b)'] = val * in_billions
         # OASDI+HI payroll tax liability (including employer share)
-        val = (recs.payrolltax * recs.s006).sum()
+        val = (vdf['payrolltax'] * wghts).sum()
         odict['Payroll Taxes ($b)'] = val * in_billions
         # combined income and payroll tax liability
-        val = (recs.combined * recs.s006).sum()
+        val = (vdf['combined'] * wghts).sum()
         odict['Combined Liability ($b)'] = val * in_billions
         # number of tax units with non-positive income tax liability
-        num = (recs.s006[recs.iitax <= 0]).sum()
+        num = (wghts[vdf['iitax'] <= 0]).sum()
         odict['With Income Tax <= 0 (#m)'] = num * in_millions
         # number of tax units with non-positive combined tax liability
-        num = (recs.s006[recs.combined <= 0]).sum()
+        num = (wghts[vdf['combined'] <= 0]).sum()
         odict['With Combined Tax <= 0 (#m)'] = num * in_millions
         return odict
     # tabulate diagnostic table
-    odict = diagnostic_table_odict(calc.records)
-    pdf = pd.DataFrame(data=odict,
-                       index=[calc.current_year],
-                       columns=odict.keys())
+    odict = diagnostic_table_odict(vdf)
+    pdf = pd.DataFrame(data=odict, index=[year], columns=odict.keys())
     pdf = pdf.transpose()
     pd.options.display.float_format = '{:8,.1f}'.format
     return pdf
-
-
-def multiyear_diagnostic_table(calc, num_years=0):
-    """
-    Generate multi-year diagnostic table from specified Calculator object.
-    This function leaves the specified calc object unchanged.
-
-    Parameters
-    ----------
-    calc : Calculator class object
-
-    num_years : integer (must be between 1 and number of available calc years)
-
-    Returns
-    -------
-    Pandas DataFrame object containing the multi-year diagnostic table
-    """
-    if num_years < 1:
-        msg = 'num_year={} is less than one'.format(num_years)
-        raise ValueError(msg)
-    max_num_years = calc.policy.end_year - calc.policy.current_year + 1
-    if num_years > max_num_years:
-        msg = ('num_year={} is greater '
-               'than max_num_years={}').format(num_years, max_num_years)
-        raise ValueError(msg)
-    cal = copy.deepcopy(calc)
-    dtlist = list()
-    for iyr in range(1, num_years + 1):
-        cal.calc_all()
-        dtlist.append(create_diagnostic_table(cal))
-        if iyr < num_years:
-            cal.increment_year()
-    return pd.concat(dtlist, axis=1)
 
 
 def mtr_graph_data(calc1, calc2,
@@ -1377,30 +1347,25 @@ def bootstrap_se_ci(data, seed, num_samples, statistic, alpha):
     return bsest
 
 
-def dec_graph_data(calc1, calc2):
+def dec_graph_data(vdf1, vdf2, year):
     """
     Prepare data needed by dec_graph_plot utility function.
 
     Parameters
     ----------
-    calc1 : a Calculator object that refers to baseline policy
+    vdf1 : a Pandas DataFrame object that refers to baseline policy and
+           contains variable in the DIFF_VARIABLES list
 
-    calc2 : a Calculator object that refers to reform policy
+    vdf2 : a Pandas DataFrame object that refers to reform policy and
+           contains variable in the DIFF_VARIABLES list
+
+    year : calendar year from which the vdf1 and vdf2 data are drawn
 
     Returns
     -------
     dictionary object suitable for passing to dec_graph_plot utility function
     """
-    # check that two calculator objects have the same current_year
-    if calc1.current_year == calc2.current_year:
-        year = calc1.current_year
-    else:
-        msg = 'calc1.current_year={} != calc2.current_year={}'
-        raise ValueError(msg.format(calc1.current_year, calc2.current_year))
-    # create difference table from the two Calculator objects
-    calc1.calc_all()
-    calc2.calc_all()
-    diff_table = create_difference_table(calc1, calc2,
+    diff_table = create_difference_table(vdf1, vdf2,
                                          groupby='weighted_deciles',
                                          income_measure='expanded_income',
                                          tax_to_diff='combined')