+++ RELEASES.md
@@ -10,14 +10,17 @@ for a complete commit history.
 [#xxxx](https://github.com/open-source-economics/Tax-Calculator/pull/xxxx))
 
 **API Changes**
-- Remove calculation of AGI tables from the TaxBrain Interface, tbi
-  [[#1724](https://github.com/open-source-economics/Tax-Calculator/pull/1724)
-  by Martin Holmer as suggested by Matt Jensen and Hank Doupe]
+- Change arguments for the three create-table utility functions and add easier to use Calculator methods that generate results tables
+  [[#1718](https://github.com/open-source-economics/Tax-Calculator/pull/1718)
+  by Martin Holmer]
 
 **New Features**
 - Add new policy parameter for refunding the new CTC against all payroll taxes
   [[#1716](https://github.com/open-source-economics/Tax-Calculator/pull/1716)
   by Matt Jensen as suggested by Ernie Tedeschi]
+- Remove calculation of AGI tables from the TaxBrain Interface, tbi
+  [[#1724](https://github.com/open-source-economics/Tax-Calculator/pull/1724)
+  by Martin Holmer as suggested by Matt Jensen and Hank Doupe]
 
 **Bug Fixes**
 - Fix Behavior.response method to handle very high marginal tax rates